pyproject.toml
[project]
name = "raytracing"
version = "0.1.0"
description = "A Python Path Tracer with C++ Core"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.0.0",
    "pillow>=10.0.0",
    "oidn",
    "nanobind",
    "imageio[freeimage,opencv,pyav]>=2.37.2",
    "tqdm>=4.67.1",
    "imageio-ffmpeg>=0.6.0",
]

[build-system]
requires = ["scikit-build-core>=0.3.3", "nanobind>=1.3.2"]
build-backend = "scikit_build_core.build"

[tool.scikit-build]
cmake.version = ">=3.15"
